+++ src/class/index.ts
@@ -3,3 +3,6 @@ export * from './FileBuilder';
 export * from './JSONDatabase';
 export * from './StringSelectMenuPaginatorBuilder';
 export * from './BoostDetector';
+export {
+    ButtonsPaginatorBuilder
+} from './ButtonsPaginatorBuilder';