+++ src/error/index.ts
@@ -9,6 +9,10 @@ export const errorkeys = {
     'TwoParamsChosen': 'TWO_PARAMS_CHOSEN',
 
     'NodeVersionError': 'NODE_VERSION_ERR',
+
+    'LengthMaxReached': 'LENGTH_MAX_REACHED',
+
+    'InputNotAllowed': 'INPUT_NOT_ALLOWED'
 };
 
 const messages = {
@@ -19,9 +23,13 @@ const messages = {
     [errorkeys.TypeError]: (type: string) => 'The parameter is not type of ' + type + '.',
     [errorkeys.MissingParam]: () => 'A required parameter is not provided.',
 
-    [errorkeys.TwoParamsChosen]: () => 'Two parameters has been set to true, expect one to be true.',
+    [errorkeys.TwoParamsChosen]: () => 'Two parameters has been set to an undefined value, expect one to be undefined value.',
 
-    [errorkeys.NodeVersionError]: (versionRequired: string) => 'Your Node.js version is currently under the required version ' + versionRequired + '. Please install a version higher than the installed one.'
+    [errorkeys.NodeVersionError]: (versionRequired: string) => 'Your Node.js version is currently under the required version ' + versionRequired + '. Please install a version higher than the installed one.',
+
+    [errorkeys.LengthMaxReached]: (variableandlength: string) => `The variable/property ${variableandlength[0]} has reached the max length of ${variableandlength[1]}.`,
+
+    [errorkeys.InputNotAllowed]: (between: any[]) => `The variable/property has received an input that is not in [${between.map((b) => b).join(', ')}]`
 };
 
 export class DJSError extends Error {